llm-worker/requirements.txt
google-genai>=1.31.0
redis>=6.4.0
python-dotenv>=1.1.0